ZMQ: extension du num dans la structure SET_BIT.
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@3993 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ trunk/Watchdogd/Include/Zmq.h
@@ -52,8 +52,8 @@
   };
 
  struct ZMQ_SET_BIT
-  { gint8 type;
-    gint8 num;
+  { gint type;
+    gint num;
     gchar acronyme [ NBR_CARAC_ACRONYME_MNEMONIQUE_UTF8+1 ];
     gchar dls_tech_id [ NBR_CARAC_PLUGIN_DLS_TECHID ];
   };
